New feature - Don't allow negative values
This change prevents the display of ERROR on the LCD module. Repeatedly pressing button2 will decrement the value until the minimum valid value is reached. At that point the value will not change when an attempt is made to decrement it.

Corrected argument type for setDisplayToDecNumber(). This method worked before but this version is more correct.